pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plantdisease"
version = "0.1.0"
description = "Plant disease detection system using deep learning"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "pillow>=8.0.0",
    "scikit-image>=0.18.0",
    "scikit-learn>=1.0.0",
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=22.0.0", "flake8>=4.0.0"]
